% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{WriteSnpPairs_sym}
\alias{WriteSnpPairs_sym}
\title{Write out epistasis result (symmetric matrix)}
\usage{
WriteSnpPairs_sym(Zmatrix, indexArr, outfile = "NONE", ...)
}
\arguments{
\item{Zmatrix}{is the Z score matrix (symmetric matrix).}

\item{indexArr}{is the index of Zmarix whose z score is over the given \code{zpthres}.}

\item{outfile}{is the SNP pairs file for the second stage.}

\item{...}{not used.}
}
\value{
null
}
\description{
Write out the result of epistasis analysis. Z score matrix is a symmetric matrix.
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
